Unified moved_monitor_wait() for moved
