+++ src/daemon/moved.cpp
@@ -43,6 +43,12 @@
 
 #include "psmove.h"
 
+#if defined(__APPLE__)
+#  include <sys/poll.h>
+#elif defined(__linux)
+#  include <poll.h>
+#endif
+
 #include <vector>
 
 struct move_daemon;
@@ -167,6 +173,9 @@ main(int argc, char *argv[])
 #else
     while (true) {
         moved.handle_request();
+        if (moved_monitor_wait(monitor, false)) {
+            moved_monitor_poll(monitor);
+        }
         moved.write_reports();
     }
 #endif
