+++ src/psmove_port.h
@@ -35,13 +35,11 @@
 #  include <unistd.h>
 #  include <sys/syslimits.h>
 #  include <sys/stat.h>
-#  include <sys/poll.h>
 #endif
 
 #ifdef __linux
 #  include <unistd.h>
 #  include <linux/limits.h>
-#  include <poll.h>
 #endif
 
 #ifdef _WIN32
